% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCtest.R
\name{VCtest}
\alias{VCtest}
\title{Test association between a SNP-set
and a zero-inflated phenotype using kernel regression.}
\usage{
VCtest(Z, obj, kernel = "linear.weighted", method = "davies",
  weights.beta = c(1, 25), weights = NULL, impute.method = "fixed",
  r.corr = 0, is_check_genotype = TRUE, is_dosage = FALSE,
  missing_cutoff = 0.15, max_maf = 1, estimate_MAF = 1)
}
\arguments{
\item{Z}{a numeric genotype matrix with each row as a different individual
and each column as a separate gene/snp. Each genotype should be coded as 0,
1, 2, and 9 (or NA) for AA, Aa, aa, and missing, where A is a major allele
and a is a minor allele. Missing genotypes will be imputed by the simple
Hardy-Weinberg equilibrium (HWE) based imputation.}

\item{obj}{an output object of the SKAT_Null_Model function.}

\item{kernel}{a type of kernel (default= "linear.weighted"). See detail
section.}

\item{method}{a method to compute the p-value (default= "davies"). "davies"
represents an exact method that computes the p-value by inverting the
characteristic function of the mixture chisq, "liu" represents an
approximation method that matches the first 3 moments, "liu.mod" represents
modified "liu" method that matches kurtosis instead of skewness to improve
tail probability approximation,"SKATO" and "optimal.adj" represent a SKAT-O
based on an unified approach, and "optimal" is an old version of the
implementation of SKAT-O. See details.}

\item{weights.beta}{a numeric vector of parameters for the beta weights for
the weighted kernels. If you want to use your own weights, please use the
"weights" parameter. It will be ignored if "weights" parameter is not null.}

\item{weights}{a numeric vector of weights for the weighted kernels. It is
\eqn{\sqrt{w}} in the SKAT paper. So if you want to use the Madsen and Browning
(2009) weight, you should set each element of weights as \eqn{1/\sqrt{p(1-p)}}, not
1/ p(1-p). When it is NULL, the beta weight with the "weights.beta"
parameter is used.}

\item{impute.method}{a method to impute missing genotypes (default= "fixed").
"bestguess" imputes missing genotypes as most likely values (0,1,2),
"random" imputes missing genotypes by generating binomial(2,p) random
variables (p is the MAF), and "fixed" imputes missing genotypes by
assigning the mean genotype values (2p).}

\item{r.corr}{the \eqn{\rho} parameter for the compound symmetric correlation
structure kernels (default= 0). If you give a vector value, SKAT will
conduct the optimal test. It will be ignored if method="optimal" or
method="optimal.adj". See details.}

\item{is_check_genotype}{a logical value indicating whether to check the
validity of the genotype matrix Z (default= TRUE). If Z has non-SNP data,
please set it FALSE, otherwise you will get an error message. If it is
FALSE and you use weighted kernels, the weights should be given through the
"weights" parameter.}

\item{is_dosage}{a logical value indicating whether the matrix Z is a dosage
matrix. If it is TRUE, SKAT will ignore "is_check_genotype".}

\item{missing_cutoff}{a cutoff of the missing rates of SNPs (default=0.15).
Any SNPs with missing rates higher than the cutoff will be excluded from
the analysis.}

\item{max_maf}{a cutoff of the maximum minor allele frequencies (MAF)
(default=1, no cutoff). Any SNPs with MAF > cutoff will be excluded from
the analysis.}

\item{estimate_MAF}{a numeric value indicating how to estimate MAFs for the
weight calculation and the missing genotype imputation. If estimate_MAF=1
(default), SKAT uses all samples to estimate MAFs. If estimate_MAF=2, only
samples with non-missing phenotypes and covariates are used to estimate
MAFs.}

\item{SSD.INFO}{an SSD_INFO object returned from Open_SSD.}

\item{SetID}{a character value of Set ID. A set ID of each set can be found
from SetInfo object in SSD.INFO.}

\item{SetIndex}{a numeric value of Set index. A set index of each set can be
found from SetInfo object in SSD.INFO.}

\item{obj.SNPWeight}{an output object of Read_SNP_WeightFile (default=NULL).
If NULL, the beta weight with the "weights.beta" parameter will be used.}
}
\value{
A list containing values from among the following (depending on
  class(obj), where obj is the output of the Null_Model function)
\item{p.val.pi}{ p-value from a test of null genetic effect on the excess
  zero parameter in the ZIP or ZINB model}
\item{p.val.ld}{  p-value from a test of null genetic effect on the mean
  parameter in the ZIP or ZINB model}
\item{p.val.cmb}{  p-value for the inverse-variance combined test of the
  global null for both the pi and lambda parameters}
\item{p.val.pi.resamp}{  resampled p-values for the pi parameter}
\item{p.val.ld.resamp}{  resampled p-values for the lambda parameter}
\item{p.val.cmb.resamp}{  resampled p-values for the combined test}
\item{p.val.Fsr}{  p-value for the Fisher method using the resampled p-value
  distributions}
\item{p.val.min}{  p-value for the min-p method using the resampled p-value
  distributions}
\item{Test.Type}{  parameter indicating the Test.Type input}
\item{Q.pi}{  the test statistic for pi}
\item{Q.ld}{  the test statistic for lambda}
\item{Q.resampling.pi}{  the resampled test statistics for pi}
\item{Q.resampling.ld}{  the resampled test statistics for lambda}
\item{param}{  some parameters from the CompQuadForm calculation which obtains
  the p-value from the distribution of the quadratic form.}
\item{p.C}{  p.value test of null genetic effect on mean parameter of non-zero
  continuous outcome in hurdle model}
\item{p.D.Adj}{  p.value for test of null genetic effect on dichotomized zero
  vs non-zero outcome in hurdle model}
\item{p.Fsr.rsmp}{  p-value from Fisher method of combining p.C and p.D, using
  resampling}
\item{p.Fsr.indp}{  p-value from Fisher method of combining p.C and p.D,
  assuming independence}
\item{p.Brown.P_test}{  p-value from Brown's method of combining p.C and p.D
  using correlation of test statistics}
\item{p.Brown.P_Fisher}{  Fisher's p-value under independence as caluculated
  by Brown's method software}
\item{p.Brown.Scale_Factor_C}{  correction factor in Brown's method}
\item{p.Brown.DF}{  degrees of freedom after correcction in Brown's method}
}
\description{
Use pre-computed null model to test association between a given set of SNPs/genes and a zero-inflated phenotype.
}
\seealso{
\code{\link{ziVCtest_Null_Model}}; \code{\link[SKAT]{SKAT}}
}
