% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziVCtest_Null_Model.R
\name{ziVCtest_Null_Model}
\alias{ziVCtest_Null_Model}
\title{Function for fitting the Null Model (under assumption of no genetic effect).}
\usage{
ziVCtest_Null_Model(formula, formula_1 = NULL, formula_2 = NULL,
  data = NULL, out_type = "ZIP", n.Resampling = 1000,
  type.Resampling = c("perturbation", "bootstrap", "permutation"),
  wgtDstr.Resampling = c("exp", "beta", "normal"))
}
\arguments{
\item{formula}{a formula object, a symbolic description of the Null Model to be fitted.
The response is a zero-inflated variable. If desired, separate models can be set for the zero and mean responses instead (see below).}

\item{formula_1}{a formula object describing the zero model. Set both \code{formula_1} and \code{formula_2}, otherwise, specifying \code{formula} will set both models equal.}

\item{formula_2}{a formula object describing the mean model. Set both \code{formula_1} and \code{formula_2}, otherwise, specifying \code{formula} will set both models equal.}

\item{data}{a data frame containing the variables in the model.}

\item{out_type}{the outcome type. options are "ZIP" (Zero Inflated Poisson), "ZINB" (Zero Inflated Negative Binomial), "Hurdle" (Hurdle Model for zero-inflated continuous outcome).}

\item{n.Resampling}{a numeric value indicating the number of resampling iterations to use for estimating the distribution of the residuals under the Null Model (default=1000).}

\item{type.Resampling}{resampling methods "perturbation", "bootstrap", or "permutation" (default="perturbation"). Not all options are available for all models.}

\item{wgtDstr.Resampling}{the distribution of the resampling weights if the \code{type.Resampling} is set to "perturbation". Options: exponential ("exp), beta ("beta"), normal ("normal").}
}
\value{
This function returns an object that has model parameters and residuals for the Null Model of no association between genetic variables and outcome phenotypes. The resampled distribution of residuals under the null hence only needs to be calculated once and reused for each SNP-set. Association testing is then carried out using the \code{VCtest} function.
}
\description{
Compute model parameters and residuals. Obtain distribution of residuals via resampling
for downstream SNP-set testing and p-values.
}
\examples{
#Modified SKAT example
data(SKAT.example)
SKAT.example.ZI = SKAT.example

set.seed(123)
SKAT.example.ZI$y.zi = ifelse(y.b, rpois(length(SKAT.example$y.b), round(exp(SKAT.example$y.c),0)), 0)
hist(SKAT.example.ZI$y.zi, xlim = c(0,20), breaks = 0:60, right = FALSE)

ziVCtest_null = ziVCtest_Null_Model(formula = y.zi ~ X,
                                    data = SKAT.example.ZI,
                                    out_type = "ZIP",
                                    n.Resampling = 500,
                                    type.Resampling = "perturbation",
                                    wgtDstr.Resampling = "exp"
)

result = VCtest(Z=SKAT.example.ZI$Z, obj=ziVCtest_null,
                kernel = "linear", method = "davies.default")

}
\seealso{
\code{\link{VCtest}} \code{\link[SKAT]{SKAT_Null_Model}}
}
